Ignore events b_call and b_return
:b_call and :b_return are small or ambiguous in profiling with ruby/profile,
and I think it is sometimes difficult to read the profile results.
We should use tools such as perf if we want profile results that require
very high accuracy.
ruby/profile is an accuracy that traces method calls and returns, and I think
it is a convenient library in most cases if it can be used simply by `-rprofile`
So I removed :b_call and :b_return event. And I want to use this ruby/profile.

By this change, we can prevent the user from being confused by seeing the
output of the number of calls as reported here.

https://bugs.ruby-lang.org/issues/15652

The reason why the number of calls is output as 5001 in reported sample is:
First, in the case of arithmetic operations, `opt_*` instruction is executed,
so :c_call event does not occur.

In the following example, `9999999999999999*999999999` or
 `9999999999999999+999999999` will be detected in :b_return event.
And since method_id at that time is `Object#slow_method` or
`Object#fast_method`, it results in calls 5001.

e.g.

require 'profile'

def slow_method
  5000.times do
    9999999999999999*999999999
  end
end

def fast_method
  5000.times do
    9999999999999999+999999999
  end
end

slow_method
fast_method